Update files to better directories and we are now working from staging
